separate DOM from logic in projects
